CSS and JS
Further updates to CSS styling.

JavaScript updated to ingest the different selections, previous this was intentionally set up for a single code to test functionality. The updated JS and corresponding HTML provides the framework to add additional tabs, switches and dedicated JS handling of inputs and outputs to code terminal.